Move resolved placement model into positions
